Split profile classes into their own separate files
The old profiles.py file was starting to become unwieldy. This messes up
history but I think it's worth doing, considering there will be more
profiles coming in the future
